backend/logic/vision_processor.py
import cv2
import mediapipe as mp
import numpy as np

class VisionProcessor:
    def __init__(self):
        self.mp_pose = mp.solutions.pose
        self.pose = self.mp_pose.Pose(
            static_image_mode=False,
            model_complexity=1, 
            min_detection_confidence=0.5
        )
    
    def process_frame(self, frame_data):
        """
        Process a video frame to check if arms are locked (elbow angle ~180).
        For this prototype, we simulate the result since we don't have a live camera stream.
        """
        # Simulation Logic:
        # Returns a mock angle and feedback
        
        # Real implementation would be:
        # image = cv2.imdecode(frame_data, cv2.IMREAD_COLOR)
        # results = self.pose.process(image)
        # Calculate angle between Shoulder, Elbow, Wrist
        
        return {
            "elbow_angle": 175, # Degrees, nearly straight
            "posture_feedback": "Arms Locked (Good)"
        }
