requirements.txt
fastapi
uvicorn
websockets
numpy
scipy
scikit-learn
google-generativeai
mediapipe
opencv-python-headless
